chore: split prompts and utils in distinct modules
